requirements.txt
mido==1.2.10
pyFluidSynth==1.3.3
python-multipart==0.0.12
python-rtmidi==1.5.8
gradio==3.50.2
PyYAML==6.0.1
